backend/requirements.txt
aioice==0.9.0
aiortc==1.10.1
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
av==13.1.0
billiard==4.2.1
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.6
django-channels==0.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
flower==2.0.1
geojson==3.2.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-crc32c==1.6.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
IMAPClient==3.0.1
incremental==24.7.2
kombu==5.4.2
Markdown==3.7
oauthlib==3.2.2
packaging==24.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyee==12.1.1
PyJWT==2.10.1
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
